Sim dev and WIP windows install

+++ gym_deepdrive/envs/deepdrive_env.py
@@ -74,13 +74,20 @@ class DeepDriveEnv(gym.Env):
         self.dashboard_queue = None
         self.should_exit = False
 
-        if not os.path.exists(c.SIM_BIN_PATH):
-            print('\n--------- Simulator does not exist, downloading ~1GB sim ----------')
-            download(c.SIM_BIN_URL, c.SIM_PATH, warn_existing=False, overwrite=False)
-        utils.ensure_executable(c.SIM_BIN_PATH)
+        if c.SIM_DEV:
+            # Sim opening and closing managed manually
+            self.sim_process = None
+        else:
+            if not os.path.exists(c.SIM_BIN_PATH):
+                print('\n--------- Simulator does not exist, downloading ~1GB sim ----------')
+                if c.IS_LINUX:
+                    download(c.SIM_BIN_URL, c.SIM_PATH, warn_existing=False, overwrite=False)
+                else:
+                    raise NotImplementedError('Sim download not yet implemented for this OS')
+            utils.ensure_executable(c.SIM_BIN_PATH)
 
-        log.info('Starting simulator at %s (takes a few seconds the first time).', c.SIM_BIN_PATH)
-        self.sim_process = Popen([c.SIM_BIN_PATH])
+            log.info('Starting simulator at %s (takes a few seconds the first time).', c.SIM_BIN_PATH)
+            self.sim_process = Popen([c.SIM_BIN_PATH])
 
         self.control = deepdrive_control.DeepDriveControl()
         self.reset_capture()
@@ -113,100 +120,6 @@ class DeepDriveEnv(gym.Env):
             # TODO: Deal with plot UI not being in the main thread somehow - (move to browser?)
             log.warning('Dashboard not supported in debug mode')
             return
-        import matplotlib.animation as animation
-        import matplotlib
-        try:
-            # noinspection PyUnresolvedReferences
-            import matplotlib.pyplot as plt
-        except ImportError as e:
-            log.error('\n\n\n***** Error: Could not start dashboard: %s\n\n', e)
-            return
-
-        def dashboard(dash_queue):
-            plt.figure(0)
-
-            class Disp(object):
-                stats = {}
-                txt_values = {}
-                lines = {}
-                x_lists = {}
-                y_lists = {}
-
-            def get_next(block=False):
-                try:
-                    q_next = dash_queue.get(block=block)
-                    if q_next['should_stop']:
-                        print('Stopping dashboard')
-                        try:
-                            anim._fig.canvas._tkcanvas.master.quit()  # Hack to avoid "Exiting Abnormally"
-                        finally:
-                            exit()
-                    else:
-                        Disp.stats = q_next['display_stats']
-                except queue.Empty:
-                    # Reuuse old stats
-                    pass
-
-            get_next(block=True)
-
-            font = {'size': 8}
-
-            matplotlib.rc('font', **font)
-
-            for i, (stat_name, stat) in enumerate(Disp.stats.items()):
-                stat = Disp.stats[stat_name]
-                stat_label_subplot = plt.subplot2grid((len(Disp.stats), 3), (i, 0))
-                stat_value_subplot = plt.subplot2grid((len(Disp.stats), 3), (i, 1))
-                stat_graph_subplot = plt.subplot2grid((len(Disp.stats), 3), (i, 2))
-                stat_label_subplot.text(0.5, 0.5, stat_name, fontsize=12, va="center", ha="center")
-                txt_value = stat_value_subplot.text(0.5, 0.5, '', fontsize=12, va="center", ha="center")
-                Disp.txt_values[stat_name] = txt_value
-                stat_graph_subplot.set_xlim([0, 200])
-                stat_graph_subplot.set_ylim([stat['ymin'], stat['ymax']])
-                Disp.lines[stat_name], = stat_graph_subplot.plot([], [])
-                stat_label_subplot.axis('off')
-                stat_value_subplot.axis('off')
-                Disp.x_lists[stat_name] = deque(np.linspace(200, 0, num=400))
-                Disp.y_lists[stat_name] = deque([-1] * 400)
-                frame1 = plt.gca()
-                frame1.axes.get_xaxis().set_visible(False)
-
-            plt.subplots_adjust(hspace=0.88)
-            fig = plt.gcf()
-            fig.set_size_inches(5.5, len(Disp.stats) * 0.5)
-            fig.canvas.set_window_title('Dashboard')
-            anim = None
-
-            def init():
-                lines = []
-                for s_name in Disp.stats:
-                    line = Disp.lines[s_name]
-                    line.set_data([], [])
-                    lines.append(line)
-                return lines
-
-            def animate(_i):
-                lines = []
-                get_next()
-                for s_name in Disp.stats:
-                    s = Disp.stats[s_name]
-                    xs = Disp.x_lists[s_name]
-                    ys = Disp.y_lists[s_name]
-                    tv = Disp.txt_values[s_name]
-                    line = Disp.lines[s_name]
-                    val = s['value']
-                    tv.set_text(str(round(val, 2)) + s['units'])
-                    ys.pop()
-                    ys.appendleft(val)
-                    line.set_data(xs, ys)
-                    lines.append(line)
-                plt.draw()
-                return lines
-
-            # TODO: Add blit=True and deal with updating the text if performance becomes unacceptable
-            anim = animation.FuncAnimation(fig, animate, init_func=init, frames=200, interval=100)
-            plt.show()
-
         q = Queue()
         p = Process(target=dashboard, args=(q,))
         p.start()
@@ -431,7 +344,8 @@ class DeepDriveEnv(gym.Env):
         deepdrive_control.close()
         if self.sess:
             self.sess.close()
-        self.sim_process.kill()
+        if self.sim_process is not None:
+            self.sim_process.kill()
 
     def _render(self, mode='human', close=False):
         # TODO: Implement proper render - this is really only good for one frame - Could use our OpenGLUT viewer (on raw images) for this or PyGame on preprocessed images
@@ -630,3 +544,96 @@ class DeepDriveRewardCalculator(object):
         progress_reward *= balance_coeff
         progress_reward = DeepDriveRewardCalculator.clip(progress_reward)
         return progress_reward
+
+def dashboard(dash_queue):
+    import matplotlib.animation as animation
+    import matplotlib
+    try:
+        # noinspection PyUnresolvedReferences
+        import matplotlib.pyplot as plt
+    except ImportError as e:
+        log.error('\n\n\n***** Error: Could not start dashboard: %s\n\n', e)
+        return
+    plt.figure(0)
+
+    class Disp(object):
+        stats = {}
+        txt_values = {}
+        lines = {}
+        x_lists = {}
+        y_lists = {}
+
+    def get_next(block=False):
+        try:
+            q_next = dash_queue.get(block=block)
+            if q_next['should_stop']:
+                print('Stopping dashboard')
+                try:
+                    anim._fig.canvas._tkcanvas.master.quit()  # Hack to avoid "Exiting Abnormally"
+                finally:
+                    exit()
+            else:
+                Disp.stats = q_next['display_stats']
+        except queue.Empty:
+            # Reuuse old stats
+            pass
+
+    get_next(block=True)
+
+    font = {'size': 8}
+
+    matplotlib.rc('font', **font)
+
+    for i, (stat_name, stat) in enumerate(Disp.stats.items()):
+        stat = Disp.stats[stat_name]
+        stat_label_subplot = plt.subplot2grid((len(Disp.stats), 3), (i, 0))
+        stat_value_subplot = plt.subplot2grid((len(Disp.stats), 3), (i, 1))
+        stat_graph_subplot = plt.subplot2grid((len(Disp.stats), 3), (i, 2))
+        stat_label_subplot.text(0.5, 0.5, stat_name, fontsize=12, va="center", ha="center")
+        txt_value = stat_value_subplot.text(0.5, 0.5, '', fontsize=12, va="center", ha="center")
+        Disp.txt_values[stat_name] = txt_value
+        stat_graph_subplot.set_xlim([0, 200])
+        stat_graph_subplot.set_ylim([stat['ymin'], stat['ymax']])
+        Disp.lines[stat_name], = stat_graph_subplot.plot([], [])
+        stat_label_subplot.axis('off')
+        stat_value_subplot.axis('off')
+        Disp.x_lists[stat_name] = deque(np.linspace(200, 0, num=400))
+        Disp.y_lists[stat_name] = deque([-1] * 400)
+        frame1 = plt.gca()
+        frame1.axes.get_xaxis().set_visible(False)
+
+    plt.subplots_adjust(hspace=0.88)
+    fig = plt.gcf()
+    fig.set_size_inches(5.5, len(Disp.stats) * 0.5)
+    fig.canvas.set_window_title('Dashboard')
+    anim = None
+
+    def init():
+        lines = []
+        for s_name in Disp.stats:
+            line = Disp.lines[s_name]
+            line.set_data([], [])
+            lines.append(line)
+        return lines
+
+    def animate(_i):
+        lines = []
+        get_next()
+        for s_name in Disp.stats:
+            s = Disp.stats[s_name]
+            xs = Disp.x_lists[s_name]
+            ys = Disp.y_lists[s_name]
+            tv = Disp.txt_values[s_name]
+            line = Disp.lines[s_name]
+            val = s['value']
+            tv.set_text(str(round(val, 2)) + s['units'])
+            ys.pop()
+            ys.appendleft(val)
+            line.set_data(xs, ys)
+            lines.append(line)
+        plt.draw()
+        return lines
+
+    # TODO: Add blit=True and deal with updating the text if performance becomes unacceptable
+    anim = animation.FuncAnimation(fig, animate, init_func=init, frames=200, interval=100)
+    plt.show()
